requirements.txt
flask==3.0.0
trakt.py==4.4.0
apscheduler==3.10.4
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
selenium==4.16.0
pyautogui==0.9.54
